Complete refactor of all remaining tests: organize into folders and add tags Simplify workaround for clicking the button to enable mushaf view
