+++ src/components/QuranReader/ContextMenu/components/MobileReadingTabs.tsx
@@ -159,6 +159,7 @@ const MobileReadingTabs: React.FC<MobileReadingTabsProps> = ({ t }) => {
             role="tab"
             tabIndex={0}
             id={tab.id}
+            data-testid={tab.id}
             onClick={() => onViewSwitched(tab.value as ReadingPreference)}
             onKeyDown={(event) => {
               if (event.key === 'Enter' || event.key === ' ') {
