+++ tests/POM/mushaf-mode.ts
@@ -11,12 +11,10 @@ export default class QuranPage {
     if (isMobile) {
       // scroll down a little to make the tab visible (bypassing a render issue)
       // FIXME: Remove this workaround when the underlying issue is fixed
-      await this.page.keyboard.press('ArrowDown', { delay: 100 });
-      await this.page.keyboard.press('ArrowDown', { delay: 100 });
-      await this.page.keyboard.press('ArrowDown', { delay: 100 });
-      await this.page.keyboard.press('ArrowUp', { delay: 100 });
+      await this.page.mouse.wheel(0, 200);
+      await this.page.mouse.wheel(0, -100);
 
-      await this.page.locator('#reading-tab').click();
+      await this.page.getByTestId('reading-tab').click();
     } else {
       await this.page.getByTestId('reading-button').click();
     }
