+++ tests/integration/content-display-and-reading/mushaf-mode/basmala-visibility.spec.ts
@@ -1,0 +1,50 @@
+import { test, expect } from '@playwright/test';
+
+import QuranPage from '@/tests/POM/mushaf-mode';
+
+test(
+  'Basmala is not shown for surahs 1 and 9 in translation mode',
+  { tag: ['@fast', '@mushaf', '@basmala'] },
+  async ({ page }) => {
+    await page.goto('/1', { waitUntil: 'networkidle' });
+
+    await expect(page.getByTestId('bismillah-section')).not.toBeVisible();
+
+    await page.goto('/9', { waitUntil: 'networkidle' });
+
+    await expect(page.getByTestId('bismillah-section')).not.toBeVisible();
+  },
+);
+
+test(
+  'Basmala is shown for other surahs in translation mode',
+  { tag: ['@fast', '@mushaf', '@basmala'] },
+  async ({ page }) => {
+    await page.goto('/2', { waitUntil: 'networkidle' });
+
+    await expect(page.getByTestId('bismillah-section')).toBeVisible();
+  },
+);
+
+test(
+  'Basmala display rules in mushaf mode - shown for every surah except 1 and 9',
+  { tag: ['@slow', '@mushaf', '@basmala'] },
+  async ({ page, isMobile }) => {
+    // Surah 76 - has basmala
+    await page.goto('/76', { waitUntil: 'networkidle' });
+
+    // Enable mushaf mode
+    const qp = new QuranPage(page);
+    await qp.mushafMode(isMobile);
+
+    await expect(page.getByTestId('bismillah-section')).toBeVisible();
+
+    // Surah 1 - no basmala
+    await page.goto('/1', { waitUntil: 'networkidle' });
+    await expect(page.getByTestId('bismillah-section')).not.toBeVisible();
+
+    // Go to Surah 9 - no basmala
+    await page.goto('/9', { waitUntil: 'networkidle' });
+    await expect(page.getByTestId('bismillah-section')).not.toBeVisible();
+  },
+);
